Feat: create mainwindow_view
